signed/unsigned c++ compiler warning fixes

+++ include/torch-mlir/Conversion/TorchOnnxToTorch/Patterns.h
@@ -60,7 +60,7 @@ struct OpBinder {
                              int64_t numOperands) {
     if (op->getNumOperands() != numOperands)
       return failure();
-    for (int i = 0; i < numOperands; i++) {
+    for (int64_t i = 0; i < numOperands; i++) {
       Value curr = op->getOperand(i);
       if (!toValidTensorType(curr.getType())) {
         return failure();
@@ -80,7 +80,7 @@ struct OpBinder {
   }
   
   ParseResult tensorOperandsList( llvm::SmallVectorImpl<Value> &values) {
-    for (int i = 0; i < op->getNumOperands(); i++) {
+    for (uint32_t i = 0; i < op->getNumOperands(); i++) {
       values.push_back(op->getOperand(i));
     }
     return success();
